Force OpenMP to use not more threads than count_sched_cpus.
Setting the number of cores with taskset does not modify the value of
omp_get_max_threads. and somehow setting the environment variable Does
not help either (probably because the variable is not exported to the
remote nodes).

So, the simplest approach is to do it internally in the C side during
the initialization.